requirements.txt
matplotlib==3.10.1
mlx==0.25.0
mlx_lm==0.23.2
mlx_optimizers==0.4.1
numpy==2.2.5
PyYAML==6.0.2
tokenizers==0.21.1
tqdm==4.67.1
# Optional for distributed CUDA support
torch>=2.0.0
